requirements.txt
matplotlib==3.2.1
pytz==2019.3
telegram==0.0.1
beautifulsoup4==4.8.2
Pillow==7.0.0
